Update patch set 4
Patch Set 4:

Hi Folks,

Sorry I just became aware of this today.

So we already setup protoc in /usr/local/bin/protoc and the acumos python client project has been successfully using it since early January. This tells me that this build should be able to access protoc if it's looking in /usr/local/bin/protoc for the package.

From the logs it looks like this project is trying to call protoc from a hardcoded path:

/w/workspace/generic-model-runner-maven-dl-verify-master-mvn33-openjdk8/bin/protoc

For this patch to work the code needs to not using the hardcoded path and ideally just call "protoc" without a path to pick up the default OR if we want to hardcode for some reason "/usr/local/bin/protoc"

Hope this helps,
Thanh

Patch-set: 4